server: progress calculation change

+++ internal/model/progress.go
@@ -1,7 +1,7 @@
 package model
 
 type Progress struct {
-	Progress          float64 `json:"progress"`
-	CountCompleted    int     `json:"count_completed"`
-	CountNotCompleted int     `json:"count_not_completed"`
+	Progress          int `json:"progress"`
+	CountCompleted    int `json:"count_completed"`
+	CountNotCompleted int `json:"count_not_completed"`
 }
